+++ packages/utils/src/fixtures.ts
@@ -1,6 +1,19 @@
 import { microtime } from './isomorphic.js';
 import { TLengthArray } from './types.js';
 
+export const randomString = (length: number = 10) => {
+	const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
+	let result = '';
+	for (let i = 0; i < length; i++) {
+		result += chars.charAt(Math.floor(Math.random() * chars.length));
+	}
+	return result;
+};
+
+export const randomInteger = (min: number = 1, max: number = 100): number => Math.floor(Math.random() * 20000) - 10000;
+
+export const randomDecimal = (min: number = -10000, max: number = 10000) => Math.random() * (max - min) + min;
+
 export function sampleRange(min = 1, max = 100) {
 	return min + Math.floor(Math.random() * (max - min + 1));
 }
