+++ packages/yantrix-parser/src/fixtures/expressions.ts
@@ -1,6 +1,6 @@
 import { Map } from 'immutable';
 import { ExpressionTypes } from '../types/expressions.js';
-import { randomString, randomInteger, randomDecimal } from '../utils/utils.js';
+import { randomString, randomInteger, randomDecimal } from '@yantrix/utils';
 
 const getStringExpressionProperties = (value: string = randomString()) => ({
 	StringDeclaration: value,
