+++ packages/yantrix-parser/src/utils/utils.ts
@@ -1,17 +1,5 @@
 import { expressionProperties } from '../fixtures/expressions.js';
-
-export const randomString = (length: number = 10) => {
-	const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
-	let result = '';
-	for (let i = 0; i < length; i++) {
-		result += chars.charAt(Math.floor(Math.random() * chars.length));
-	}
-	return result;
-};
-
-export const randomInteger = (min: number = 1, max: number = 100): number => Math.floor(Math.random() * 20000) - 10000;
-
-export const randomDecimal = (min: number = -10000, max: number = 10000) => Math.random() * (max - min) + min;
+import { randomString, randomInteger, randomDecimal } from '@yantrix/utils';
 
 const KeyItemsCount = 50;
